+++ docs/CHANGELOG.md
@@ -5,3 +5,4 @@
 + Use the next generation overlay
 + Add grid size precision to snap intersection, difference and symdifference operators
 + Improve ST_Union to use CoverageUnion when the input geometry contains only polygons
++ Use OverlayNGRobust instead of OverlayNG